seved until

+++ src/models/Role.js
@@ -2,7 +2,12 @@ import { DataTypes } from "sequelize";
 import { sequelize } from "../config/db.js";
 
 const Role = sequelize.define("Role", {
-  id: { type: DataTypes.INTEGER.UNSIGNED, autoIncrement: true, primaryKey: true },
+id: {
+  type: DataTypes.BIGINT.UNSIGNED,
+  autoIncrement: true,
+  primaryKey: true
+}
+,
   name: { type: DataTypes.STRING(50), allowNull: false, unique: true },
   description: { type: DataTypes.STRING(255) }
 }, { tableName: "roles" });
